WCP-84 : Created IT test cases for session interaction

+++ blackboardvc-portlet-api/src/main/java/org/jasig/portlet/blackboardvcportlet/dao/ws/SessionWSDao.java
@@ -14,25 +14,25 @@ public interface SessionWSDao {
 	
 	//create
 	public BlackboardSessionResponse createSession(ConferenceUser user, SessionForm sessionForm, boolean fullAccess);
-	public String buildSessionUrl(long sessionId, String displayName);
+	public String buildSessionUrl(Long sessionId, String displayName);
 	//TODO : this might just go into create session
-	public boolean createSessionTelephony(int sessionId, BlackboardSetSessionTelephony telephony);
+	public boolean createSessionTelephony(Long sessionId, BlackboardSetSessionTelephony telephony);
 
 	//read
 	//TODO : Determine if this is needed as we should be reading from the local database.
 	public List<BlackboardSessionResponse> getSessions(String userId, String groupingId, Long sessionId, String creatorId, Long startTime, Long endTime, String sessionName);
 	
 	//Misc. reads that have to do with a session but don't warrant there own dao
-	public List<BlackboardSessionAttendanceResponse> getSessionAttendance(int sessionId, Object startTime);
-	public List<BlackboardSessionTelephonyResponse> getSessionTelephony(int sessionId);
+	public List<BlackboardSessionAttendanceResponse> getSessionAttendance(Long sessionId, Object startTime);
+	public List<BlackboardSessionTelephonyResponse> getSessionTelephony(Long sessionId);
 	
 	//update
 	public BlackboardSessionResponse updateSession(ConferenceUser user, SessionForm sessionForm, boolean fullAccess);
 	
 	//delete
-	public boolean deleteSession(int sessionId);
-	public boolean clearSessionChairList(int sessionId);
-	public boolean clearSessionNonChairList(int sessionId);
+	public boolean deleteSession(Long sessionId);
+	public boolean clearSessionChairList(Long sessionId);
+	public boolean clearSessionNonChairList(Long sessionId);
